ns8-join. Show the VPN endpoint address
Store in e-smith ns8 record the VPN endpoint instead of the connect
address. This is more important to identify the cluster and to
understand that the VPN connection is the channel that is used for the
whole migration process.

On the contrary, the connection address is used just to execute the
add-node action.